requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-dev --no-hashes --frozen --output-file=requirements.txt
-e .
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.0
    # via assorted-hooks
aiosignal==1.4.0
    # via aiohttp
attrs==25.4.0
    # via aiohttp
certifi==2025.10.5
    # via requests
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.4.4
    # via requests
cryptography==46.0.3
    # via pyjwt
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.11
    # via
    #   requests
    #   yarl
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
packaging==25.0
    # via assorted-hooks
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pycparser==2.23 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pygithub==2.8.1
    # via assorted-hooks
pyjwt==2.10.1
    # via pygithub
pynacl==1.6.0
    # via pygithub
pyyaml==6.0.3
    # via assorted-hooks
requests==2.32.5
    # via pygithub
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   assorted-hooks
    #   pygithub
urllib3==2.5.0
    # via
    #   pygithub
    #   requests
yarl==1.22.0
    # via aiohttp
